+++ lib/src/injection_context.dart
@@ -14,6 +14,8 @@
 // limitations under the License.
 //
 
+import 'dart:collection';
+
 import 'exceptions.dart';
 
 String _stringOfType<T>() => T.toString();
@@ -24,7 +26,7 @@ typedef ServiceInitializer<T> = T Function();
 
 /// This the function type definition for the initialization of the available
 /// services.
-typedef InjectionInitializer = void Function();
+typedef InjectionInitializer = void Function(Context);
 
 class _ServiceConfiguration<T> {
   String name;
@@ -35,28 +37,101 @@ class _ServiceConfiguration<T> {
   _ServiceConfiguration({this.name, this.serviceInitializer, this.singleton});
 }
 
+/// This singleton class holds all context for the different profiles.
+class _ContextCollection {
+  /// Name of the global profiles.
+  static final String globalProfile = '_____GLOBAL_____';
+
+  /// The singleton instance.
+  static final _ContextCollection _singleton = _ContextCollection._global();
+
+  /// Maps with all assignments from profile name to its injection context.
+  Map<String, _InjectionContext> profiles = {
+    _ContextCollection.globalProfile: _InjectionContext(profile: _ContextCollection.globalProfile)
+  };
+
+  /// All active profiles
+  List<String> activeProfiles = [];
+
+  /// Internal constructor for the singleton.
+  _ContextCollection._global();
+
+  /// Getter for the shared instance.
+  static _ContextCollection get shared => _ContextCollection._singleton;
+
+  List<_InjectionContext> get _activeContexts =>
+      activeProfiles.map<_InjectionContext>((profile) => _getContext(profile)).toList();
+
+  _InjectionContext get globalContext {
+    return _getContext(globalProfile);
+  }
+
+  _InjectionContext _getContext(String profile) {
+    var context;
+
+    if (profiles.containsKey(profile)) {
+      context = profiles[profile];
+    } else {
+      context = _InjectionContext(profile: profile);
+      profiles[profile] = context;
+    }
+
+    return context;
+  }
+
+  void startupContext(String profile, InjectionInitializer initializer) {
+    _getContext(profile).startup(initializer);
+  }
+
+  void shutdown() {
+    profiles.values.forEach((context) => context.shutdown());
+    profiles = {};
+    activeProfiles = [];
+  }
+}
+
+/// This abstract class is implemented by the internal injection context
+/// and provides the interface for registering services.
+abstract class Context {
+  /// This method shall be called to register a service with the injection
+  /// context. It must only be called, after the [startup] has been called.
+  ///
+  /// The [initializer] is called when an instance of the service shall be
+  /// created. It must return the instance of the service. By passing a [name],
+  /// it is possible to register different services that implement the same
+  /// class. If the flag [asSingleton] is true, the [globalInitializer] is only called
+  /// once when resolving the service. The created service instance will be
+  /// cached. If the flag [asSingleton] is false, the [initializer] is called
+  /// every time when resolving the service.
+  ///
+  /// Throws a [InjectionContextNotInitialized] exception, if the injection
+  /// context is not started.
+  ///
+  /// Throws a [InjectionContextHasAlreadyService] exception, if the service
+  /// that shall be registered was already registered before.
+  void register<T>(
+    ServiceInitializer<T> globalInitializer, {
+    String name,
+    bool asSingleton = true,
+  });
+}
+
 /// The [_InjectionContext] is the global registry for all services. It is a
 /// singleton, since the context must be the same instance in the entire
 /// application.
-class _InjectionContext {
-  /// The singleton instance of the [_InjectionContext].
-  static final _InjectionContext _singleton = _InjectionContext._global();
+class _InjectionContext implements Context {
+  String profile;
 
   /// This flag determines whether the [startup] method has been called or not.
   /// If the [startup] method has not been calledm the [_InjectionContext] is not
   /// functional and each call to any method results in an error.
   bool _initialized = false;
 
-  Map<String, _ServiceConfiguration> _services = {};
-
-  /// Internal constructor for the singleton.
-  _InjectionContext._global();
+  Map<String, _ServiceConfiguration> _services = HashMap<String, _ServiceConfiguration>();
 
   /// This factory always returns the singleton instance of the
   /// [_InjectionContext].
-  factory _InjectionContext() {
-    return _singleton;
-  }
+  _InjectionContext({this.profile});
 
   /// This method must be called at the very beginning of the application.
   /// It calls the [InjectionInitializer] to initialize itself.
@@ -67,14 +142,14 @@ class _InjectionContext {
 
     _initialized = true;
 
-    initializer();
+    initializer(this);
   }
 
   /// This method can be called to reset the injection context. Every registered
   /// service and all singleton instances will be removed.
-  void shutDown() {
+  void shutdown() {
     _initialized = false;
-    _services = {};
+    _services = HashMap<String, _ServiceConfiguration>();
   }
 
   /// This method shall be called to register a service with the injection
@@ -83,7 +158,7 @@ class _InjectionContext {
   /// The [initializer] is called when an instance of the service shall be
   /// created. It must return the instance of the service. By passing a [name],
   /// it is possible to register different services that implement the same
-  /// class. If the flag [asSingleton] is true, the [initializer] is only called
+  /// class. If the flag [asSingleton] is true, the [globalInitializer] is only called
   /// once when resolving the service. The created service instance will be
   /// cached. If the flag [asSingleton] is false, the [initializer] is called
   /// every time when resolving the service.
@@ -93,7 +168,12 @@ class _InjectionContext {
   ///
   /// Throws a [InjectionContextHasAlreadyService] exception, if the service
   /// that shall be registered was already registered before.
-  void register<T>(ServiceInitializer<T> initializer, {String name, bool asSingleton = true}) {
+  @override
+  void register<T>(
+    ServiceInitializer<T> globalInitializer, {
+    String name,
+    bool asSingleton = true,
+  }) {
     if (!_initialized) {
       throw InjectionContextNotInitialized();
     }
@@ -103,7 +183,7 @@ class _InjectionContext {
     }
 
     _services[_key<T>(name)] =
-        _ServiceConfiguration<T>(name: name, serviceInitializer: initializer, singleton: asSingleton);
+        _ServiceConfiguration<T>(name: name, serviceInitializer: globalInitializer, singleton: asSingleton);
   }
 
   /// This method resolves a service determined by the type [T] and an optional
@@ -153,36 +233,44 @@ class _InjectionContext {
   }
 
   String _key<T>(String name) {
-    return _stringOfType<T>() + ':' + (name ?? _stringOfType<T>());
+    var key = '';
+
+    key = key + _stringOfType<T>() + ':' + (name ?? _stringOfType<T>());
+
+    return key;
+  }
+
+  bool _hasService<T>({String name}) {
+    return _services.containsKey(_key<T>(name));
   }
 }
 
 /// This function must be called at the very beginning of the application to initialize
-/// the injection context.
-void startup(InjectionInitializer initializer) => _InjectionContext().startup(initializer);
+/// the injection context. The [initializer] is the one for the gloabl profile. The
+/// [activeProfiles] is a list of all the profile names that are currently activated.
+/// When resolving a service, the global profile and the all the activated profiles are
+/// used to search a service, where the global propfile has the least priority.
+/// The map [profileInitializers] are the initializers for the differemt know profiles.
+void startup(InjectionInitializer initializer,
+    {List<String> activeProfiles, Map<String, InjectionInitializer> profileInitializers}) {
+  _ContextCollection.shared.activeProfiles.addAll(activeProfiles ?? []);
+  _ContextCollection.shared.activeProfiles.add(_ContextCollection.globalProfile);
+
+  var globalContext = _ContextCollection.shared.globalContext;
+  globalContext.startup(initializer);
+
+  if (profileInitializers != null) {
+    profileInitializers.keys.toList().forEach((profile) {
+      _ContextCollection.shared.startupContext(profile, profileInitializers[profile]);
+    });
+  }
+}
 
 /// This function can be called to reset the injection context. Every registered
 /// service and all singleton instances will be removed.
-void shutDown() => _InjectionContext().shutDown();
-
-/// This function shall be called to register a service with the injection
-/// context. It must only be called, after the [startup] has been called.
-///
-/// The [initializer] is called when an instance of the service shall be
-/// created. It must return the instance of the service. By passing a [name],
-/// it is possible to register different services that implement the same
-/// class. If the flag [asSingleton] is true, the [initializer] is only called
-/// once when resolving the service. The created service instance will be
-/// cached. If the flag [asSingleton] is false, the [initializer] is called
-/// every time when resolving the service.
-///
-/// Throws a [InjectionContextNotInitialized] exception, if the injection
-/// context is not started.
-///
-/// Throws a [InjectionContextHasAlreadyService] exception, if the service
-/// that shall be registered was already registered before.
-void register<T>(ServiceInitializer<T> initializer, {String name, bool asSingleton = true}) =>
-    _InjectionContext().register<T>(initializer, name: name, asSingleton: asSingleton);
+void shutdown() {
+  _ContextCollection.shared.shutdown();
+}
 
 /// This function resolves a service determined by the type [T] and an optional
 /// [name].
@@ -192,10 +280,29 @@ void register<T>(ServiceInitializer<T> initializer, {String name, bool asSinglet
 ///
 /// Throws a [InjectionContextHasNoService] exception, if the service
 /// that shall be resolved was not registered before.
-T resolve<T>({String name}) => _InjectionContext().resolve<T>(name: name);
+T resolve<T>({String name}) {
+  var services = _ContextCollection.shared._activeContexts
+      .where((context) => context._hasService<T>(name: name))
+      .map<T>((context) => context.resolve<T>(name: name))
+      .toList();
+
+  if (services.length > 1) {
+    throw InjectionContextHasMoreThanOneService(services);
+  } else if (services.isEmpty) {
+    throw InjectionContextHasNoService(_stringOfType<T>(), (name ?? _stringOfType<T>()));
+  }
+
+  return services.first;
+}
 
 /// This function resolves a service determined by the type [T].
 ///
 /// Throws a [InjectionContextNotInitialized] exception, if the injection
 /// context is not started.
-List<T> resolveAll<T>() => _InjectionContext().resolveAll<T>();
+List<T> resolveAll<T>() {
+  var services = List<T>();
+
+  _ContextCollection.shared.profiles.values.forEach((context) => services.addAll(context.resolveAll<T>()));
+
+  return services;
+}